Make `limit` customizable in paginated pages (#669)
Two changes. First, we rename `items` to `limit`. Since
https://ddnexus.github.io/pagy/changelog/#version-900 (we're using
Pagy 9.3.4), `items` has been renamed to `limit`. This means that
since probably a long time the custom limits didn't really work. You
can verify this:

  - https://zondersikkel.nl/beers has 20 items per page, where
  `app/controllers/beers_controller.rb` had `items: 16`.

  - https://zondersikkel.nl/chugtypes/2 has 20 items per page,
  `chugtypes_controller.rb` has `items: 10`

  - https://zondersikkel.nl/users/14 has 20 quotes per page,
  `static_pages_controller.rb` has `items: 12`

  - etc

Change 2: make limit customizable. With this change, going to
http://localhost:3000/?limit=5 shows only 5 quotes per page. The
motivation for this was that it's somewhat burdensome to search for
quotes in the past. This way you can just set `?limit=10000` and do
Ctrl+F :)